+++ app/components/top-nav.tsx
@@ -1,19 +1,11 @@
-import type { LinksFunction } from "@remix-run/node";
-import { Link } from "@remix-run/react";
-import logo_hori from "../imgs/SocialPlan-it-logo-Horizontal.png";
-import logo_sq from "../imgs/SocialPlan-it-logo-Square.png";
-export const links: LinksFunction = () => [
-  {
-    rel: "stylesheet",
-    href: " stylesUrl",
-  },
-];
+import logo_hori from '../imgs/SocialPlan-it-logo-Horizontal.png';
+import logo_sq from '../imgs/SocialPlan-it-logo-Square.png';
 
-export function TopNav() {
+export default function TopNav() {
   return (
     <>
-      <div className="mx-auto bg-gray-700">
-        <div className="mx-auto max-w-7xl px-2 sm:px-6 lg:px-8 bg-gray-700">
+      <div className="bg-gray-700">
+        <div className="mx-40 max-w-7xl px-2 sm:px-6 lg:px-8 bg-gray-700">
           <div className="relative flex h-16 items-center justify-between">
             <div className="absolute inset-y-0 left-0 flex items-center sm:hidden">
               {/* <!-- Mobile menu button--> */}
@@ -26,7 +18,6 @@ export function TopNav() {
                 <span className="sr-only">Open main menu</span>
                 {/* <!--
               Icon when menu is closed.
-
               Menu open: "hidden", Menu closed: "block"
             --> */}
                 <svg
@@ -37,11 +28,7 @@ export function TopNav() {
                   stroke="currentColor"
                   aria-hidden="true"
                 >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5"
-                  />
+                  <path strokeLinecap="round" strokeLinejoin="round" d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5" />
                 </svg>
                 {/* <!--
               Icon when menu is open.
@@ -56,49 +43,37 @@ export function TopNav() {
                   stroke="currentColor"
                   aria-hidden="true"
                 >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M6 18L18 6M6 6l12 12"
-                  />
+                  <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
                 </svg>
               </button>
             </div>
             <div className="flex flex-1 items-center justify-center sm:items-stretch sm:justify-start">
               <div className="flex flex-shrink-0 items-center">
-                <img
-                  className="block h-8 w-auto lg:hidden"
-                  src={logo_sq}
-                  alt="Plan It Social Logo!"
-                />
-                <img
-                  className="hidden h-8 w-auto lg:block"
-                  src={logo_hori}
-                  alt="Plan It Social Logo!"
-                />
+                <img className="block h-8 w-auto lg:hidden" src={logo_sq} alt="Plan It Social Logo!" />
+                <img className="hidden h-8 w-auto lg:block" src={logo_hori} alt="Plan It Social Logo!" />
               </div>
               <div className="hidden sm:ml-6 sm:block">
                 <div className="flex space-x-4">
                   {/* <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" --> */}
-                  <Link
-                    to="/"
+                  <a
+                    href="/"
                     className="bg-gray-900 text-white rounded-md px-3 py-2 text-sm font-medium"
                     aria-current="page"
                   >
                     Home
-                  </Link>
-                  <Link
-                    to="/about-us"
+                  </a>
+                  <a
+                    href="/about-us"
                     className="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium"
                   >
                     About
-                  </Link>
-                  <Link
-                    to="/"
+                  </a>
+                  <a
+                    href="/login"
                     className="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium"
                   >
                     Login
-                  </Link>
+                  </a>
                 </div>
               </div>
             </div>
@@ -143,33 +118,47 @@ export function TopNav() {
                   </button>
                 </div>
 
+                {/* <!--
+              Dropdown menu, show/hide based on menu state.
+
+              Entering: "transition ease-out duration-100"
+                From: "transform opacity-0 scale-95"
+                To: "transform opacity-100 scale-100"
+              Leaving: "transition ease-in duration-75"
+                From: "transform opacity-100 scale-100"
+                To: "transform opacity-0 scale-95"
+            --> */}
                 <div
                   className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none"
                   role="menu"
                   aria-orientation="vertical"
                   aria-labelledby="user-menu-button"
+                  tabIndex="-1"
                 >
                   {/* <!-- Active: "bg-gray-100", Not Active: "" --> */}
                   <a
-                    href="/"
+                    href="#"
                     className="block px-4 py-2 text-sm text-gray-700"
                     role="menuitem"
+                    tabIndex="-1"
                     id="user-menu-item-0"
                   >
                     Your Profile
                   </a>
                   <a
-                    href="/"
+                    href="#"
                     className="block px-4 py-2 text-sm text-gray-700"
                     role="menuitem"
+                    tabIndex="-1"
                     id="user-menu-item-1"
                   >
                     Settings
                   </a>
                   <a
-                    href="/"
+                    href="#"
                     className="block px-4 py-2 text-sm text-gray-700"
                     role="menuitem"
+                    tabIndex="-1"
                     id="user-menu-item-2"
                   >
                     Sign out
@@ -185,31 +174,31 @@ export function TopNav() {
         <div className="sm:hidden" id="mobile-menu">
           <div className="space-y-1 px-2 pb-3 pt-2">
             {/* <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" --> */}
-            <Link
-              to="/"
+            <a
+              href="#"
               className="bg-gray-900 text-white block rounded-md px-3 py-2 text-base font-medium"
               aria-current="page"
             >
               Home
-            </Link>
-            <Link
-              to="/about-us"
+            </a>
+            <a
+              href="#"
               className="text-gray-300 hover:bg-gray-700 hover:text-white block rounded-md px-3 py-2 text-base font-medium"
             >
               About
-            </Link>
-            <Link
-              to="/"
+            </a>
+            <a
+              href="#"
               className="text-gray-300 hover:bg-gray-700 hover:text-white block rounded-md px-3 py-2 text-base font-medium"
             >
               Log
-            </Link>
-            <Link
-              to="/"
+            </a>
+            <a
+              href="#"
               className="text-gray-300 hover:bg-gray-700 hover:text-white block rounded-md px-3 py-2 text-base font-medium"
             >
               Calendar
-            </Link>
+            </a>
           </div>
         </div>
       </div>
