+++ app/root.tsx
@@ -1,6 +1,6 @@
 import type { LinksFunction } from '@remix-run/node';
 import stylesheet from './styles/tailwind.css';
-import TopNav from '~/components/topNav';
+import TopNav from '~/components/top-nav';
 import { Links, LiveReload, Meta, Outlet, Scripts, ScrollRestoration } from '@remix-run/react';
 import Footer from './components/footer';
